Added files via upload
Description
In this assignment you should implement an online hotel reservation. Requirements of this project are as follows:

1.	An admin account has following functionalities:
a.	Able to add, edit, or delete rooms and their specifications
b.	Able to gets various reports about the reserved rooms for specific range of dates, floor, and room price, etc.
2.	User account:
a.	Able to book a room
b.	Able to change the booking
c.	Able to request especial services through a text request
3.	Guest account(without login):
a.	Able to browse the room availability
b.	Able to see room specifications
c.	Able to see all public information such as “contact us”, photo gallery, map, and so on. 

On due date, you should present your project in class.

 Mark Scheme for Implementation of the Project

1.	The level of completeness: 6 Marks
2.	Correct Design: 4 Marks
3.	Correct implementation: 3 Marks
4.	Presentation: 2 Marks


Deliverable
The proposal and report should be hande